refactor: namespace layout header
