+++ src/routes/steps/connect-wallet-confirm/connect-wallet-confirm.jsx
@@ -1,7 +1,6 @@
 import { useState } from "react";
 import chevron from "../../../assets/chevron.svg";
 import { Button, Layout } from "../../../components"; //button needs to be changed when avail
-import { Header } from "../../../components/layout/header";
 import { Body } from "./body";
 
 //TODO: figure out step numbering
@@ -50,18 +49,18 @@ export const ConnectWalletConfirm = ({ onBackClick, onForwardClick }) => {
       className="mb-32px"
       header={
         <>
-          <Header.Left>
+          <Layout.Header.Left>
             <Button
               buttonKind="textOnly"
               leftIcon={chevron}
               size="sm"
               onClick={onBackClick}
             />
-          </Header.Left>
-          <Header.Center>
-            <Header.Title title="Connecting wallet" />
-            <Header.SubTitle subtitle="Step 2/4" />
-          </Header.Center>
+          </Layout.Header.Left>
+          <Layout.Header.Center>
+            <Layout.Header.Title title="Connecting wallet" />
+            <Layout.Header.SubTitle subtitle="Step 2/4" />
+          </Layout.Header.Center>
         </>
       }
       body={<Body></Body>}
