+++ src/routes/steps/repeat-seed-phrase/repeat-seed-phrase.jsx
@@ -1,7 +1,6 @@
 import { useCallback, useContext, useEffect, useState } from "react";
 import chevron from "../../../assets/chevron.svg";
 import { Button, Layout } from "../../../components"; //button needs to be changed when avail
-import { Header } from "../../../components/layout/header";
 import { WizardContext } from "../../../contexts/wizard-context";
 import { Body } from "./body";
 //TODO: Update buttons when available
@@ -74,18 +73,18 @@ export function RepeatSeedPhrase({ onBackClick, onForwardClick }) {
       className="mb-32px"
       header={
         <>
-          <Header.Left>
+          <Layout.Header.Left>
             <Button
               buttonKind="textOnly"
               leftIcon={chevron}
               size="sm"
               onClick={handleBackClick}
             />
-          </Header.Left>
-          <Header.Center>
-            <Header.Title title="Sign up" />
-            <Header.SubTitle subtitle=" Step 3/4" />
-          </Header.Center>
+          </Layout.Header.Left>
+          <Layout.Header.Center>
+            <Layout.Header.Title title="Sign up" />
+            <Layout.Header.SubTitle subtitle=" Step 3/4" />
+          </Layout.Header.Center>
         </>
       }
       body={
