+++ src/routes/steps/seed-phrase/seed-phrase.jsx
@@ -1,6 +1,5 @@
 import chevron from "../../../assets/chevron.svg";
 import { Button, Layout, Modal } from "../../../components"; //button needs to be changed when avail
-import { Header } from "../../../components/layout/header";
 import { Body } from "./body";
 //TODO: Update buttons when available
 //TODO:  Figure out how to get modal to line up with layout bottom
@@ -41,18 +40,18 @@ export function SeedPhrase({
       className="mb-32px"
       header={
         <>
-          <Header.Left>
+          <Layout.Header.Left>
             <Button
               buttonKind="textOnly"
               leftIcon={chevron}
               size="sm"
               onClick={onBackClick}
             />
-          </Header.Left>
-          <Header.Center>
-            <Header.Title title="Sign up" />
-            <Header.SubTitle subtitle=" Step 2/4" />
-          </Header.Center>
+          </Layout.Header.Left>
+          <Layout.Header.Center>
+            <Layout.Header.Title title="Sign up" />
+            <Layout.Header.SubTitle subtitle=" Step 2/4" />
+          </Layout.Header.Center>
         </>
       }
       body={
